Refactored 2d export
Most of the code is now in core.  Java2dSnapshotController was factored out
of Java2dSnapshot.

The Java2dExporter dependency on Java3d Transform3d was inlined as two
functions, ortho() and perspective().
