+++ core/src/main/java/com/vzome/core/exporters2d/Java2dExporter.java
@@ -1,4 +1,4 @@
-package org.vorthmann.zome.export.java2d;
+package com.vzome.core.exporters2d;
 
 import java.awt.Color;
 import java.awt.geom.Rectangle2D;
@@ -7,7 +7,6 @@ import java.io.Writer;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.media.j3d.Transform3D;
 import javax.vecmath.Matrix4d;
 import javax.vecmath.Point3f;
 import javax.vecmath.Point4f;
@@ -26,8 +25,8 @@ import com.vzome.core.model.Strut;
 import com.vzome.core.render.Colors;
 import com.vzome.core.render.RenderedManifestation;
 import com.vzome.core.render.RenderedModel;
-import com.vzome.core.viewing.Lights;
 import com.vzome.core.viewing.Camera;
+import com.vzome.core.viewing.Lights;
 
 
 /**
@@ -42,7 +41,7 @@ public class Java2dExporter extends Exporter3d
 	private final Vector3f[] lightDirs = new Vector3f[3];
 	private final Color[] lightColors = new Color[3];
 	private final Color ambientLight;
-	private final Transform3D viewTransform;
+	private final Matrix4d viewTransform;
 
     /**
      * @param v[]
@@ -57,7 +56,7 @@ public class Java2dExporter extends Exporter3d
         this .view = view;
         Matrix4d viewMatrix = new Matrix4d();
         this .view .getViewTransform( viewMatrix, 0d );
-        this .viewTransform = new Transform3D( viewMatrix );
+        this .viewTransform = viewMatrix;
         
         for ( int i = 0; i < lightDirs.length; i++ ) {
             lightDirs[ i ] = new Vector3f();
@@ -85,6 +84,8 @@ public class Java2dExporter extends Exporter3d
         
         mSnapshot .setStrokeWidth( 0.5f );
         mSnapshot .setRect( new Rectangle2D.Float( 0f, 0f, width, height ) );
+        
+        mSnapshot .setBackgroundColor( this .getBackgroundColor() );
 
         List<Vector3f> mappedVertices = new ArrayList<>( 60 );
         for (RenderedManifestation rm : mModel) {
@@ -183,24 +184,108 @@ public class Java2dExporter extends Exporter3d
     {
         Matrix4d viewMatrix = new Matrix4d();
         this .view .getViewTransform( viewMatrix, 0d );
-        Transform3D viewTrans = new Transform3D( viewMatrix );
-        viewTrans .transform( point );
+        viewMatrix .transform( point );
         // point is now in view coordinates
         Vector4f p4 = new Vector4f( point.x, point.y, point.z, 1f );
 
-        Transform3D eyeTrans = new Transform3D();
+        Matrix4d eyeTrans;
         if ( ! view .isPerspective() ) {
 			double edge = view .getWidth() / 2;
-			eyeTrans .ortho( -edge, edge, -edge, edge, view .getNearClipDistance(), view .getFarClipDistance() );
+			eyeTrans = ortho( -edge, edge, -edge, edge, view .getNearClipDistance(), view .getFarClipDistance() );
 		}
 		else
-		    eyeTrans .perspective( view .getFieldOfView(), 1.0d, view .getNearClipDistance(), view .getFarClipDistance() );
+		    eyeTrans = perspective( view .getFieldOfView(), 1.0d, view .getNearClipDistance(), view .getFarClipDistance() );
 			// TODO - make aspect ratio track the screen window shape
         
         eyeTrans .transform( p4 );
         point .project( new Point4f( p4 ) );
     }
 
+    /**
+     * Creates a perspective projection transform that mimics a standard,
+     * camera-based,
+     * view-model.  
+     * (From javax.media.j3d.Transform3d.java)
+     * This transform maps coordinates from Eye Coordinates (EC)
+     * to Clipping Coordinates (CC).  Note that unlike the similar function
+     * in OpenGL, the clipping coordinates generated by the resulting
+     * transform are in a right-handed coordinate system
+     * (as are all other coordinate systems in Java 3D). Also note that the
+     * field of view is specified in radians.
+     * @param fovx specifies the field of view in the x direction, in radians
+     * @param aspect specifies the aspect ratio and thus the field of
+     * view in the x direction. The aspect ratio is the ratio of x to y,
+     * or width to height.
+     * @param zNear the distance to the frustum's near clipping plane.
+     * This value must be positive, (the value -zNear is the location of the
+     * near clip plane).
+     * @param zFar the distance to the frustum's far clipping plane
+     */
+    public Matrix4d perspective( double fovx, double aspect, double zNear, double zFar )
+    {
+        Matrix4d m = new Matrix4d();
+        double sine, cotangent, deltaZ;
+        double half_fov = fovx * 0.5;
+
+        deltaZ = zFar - zNear;
+        sine = Math.sin(half_fov);
+        cotangent = Math.cos(half_fov) / sine;
+
+        m.m00 = cotangent;
+        m.m11 = cotangent * aspect;
+        m.m22 = (zFar + zNear) / deltaZ;
+        m.m23 = 2.0 * zNear * zFar / deltaZ;
+        m.m32 = -1.0;
+        m.m01 = m.m02 = m.m03 = m.m10 = m.m12 = m.m13 = m.m20 =
+                m.m21 = m.m30 = m.m31 = m.m33 = 0;
+
+        return m;
+    }
+
+
+    /**
+     * Creates an orthographic projection transform that mimics a standard,
+     * camera-based,
+     * view-model.  
+     * (From javax.media.j3d.Transform3d.java)
+     * This transform maps coordinates from Eye Coordinates (EC)
+     * to Clipping Coordinates (CC).  Note that unlike the similar function
+     * in OpenGL, the clipping coordinates generated by the resulting
+     * transform are in a right-handed coordinate system
+     * (as are all other coordinate systems in Java 3D).
+     * @param left the vertical line on the left edge of the near
+     * clipping plane mapped to the left edge of the graphics window
+     * @param right the vertical line on the right edge of the near
+     * clipping plane mapped to the right edge of the graphics window
+     * @param bottom the horizontal line on the bottom edge of the near
+     * clipping plane mapped to the bottom edge of the graphics window
+     * @param top the horizontal line on the top edge of the near
+     * clipping plane mapped to the top edge of the graphics window
+     * @param near the distance to the frustum's near clipping plane
+     * (the value -near is the location of the near clip plane)
+     * @param far the distance to the frustum's far clipping plane
+     */
+    public Matrix4d ortho( double left, double right, double bottom,
+                           double top, double near, double far )
+    {
+        Matrix4d m = new Matrix4d();
+        double deltax = 1/(right - left);
+        double deltay = 1/(top - bottom);
+        double deltaz = 1/(far - near);
+
+        m.m00 = 2.0 * deltax;
+        m.m03 = -(right + left) * deltax;
+        m.m11 = 2.0 * deltay;
+        m.m13 = -(top + bottom) * deltay;
+        m.m22 = 2.0 * deltaz;
+        m.m23 = (far + near) * deltaz;
+        m.m01 = m.m02 = m.m10 = m.m12 = m.m20 =
+            m.m21 = m.m30 = m.m31 = m.m32 = 0;
+        m.m33 = 1;
+        return m;
+    }
+
+    
     private Vector3f mapCoordinates( RealVector rv, int height, int width, AlgebraicField field, Camera view )
     {
         float xscale = width/2f;
